feat(diary): display products in meal screen
The meal screen now displays a list of products that are included in a meal.
It also allows the user to navigate to the "add product" screen.
The meal header was enhanced with the start and end time of the meal.
A back button was added to the top app bar to navigate back.